README/tests: isolate README tests to new file
